requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.6
django-crispy-forms==2.1
mysqlclient==2.2.4
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
